FileEntity => IFileEntity, FileTrait => TFileEntity, fluent all
